Moved definition of template classes in header files
Removed the cpp files containing the definition of the template classes "timer" and "shared_queue" and put their definition directly together with their declaration in the header files.
This fixes the following compilation problem: when modifying the implementation of those classes, it was necessary to perform a "make clean" before recompiling, otherwise they were skipped and the old implementation (aòready compiled) was used by the linker.